# Conflicts:
#	com-platform-admin-service-contract/src/main/java/com/platform/admin/service/client/param/JdGoodsSyncParam.java
#	com-platform-admin-service-contract/src/main/java/com/platform/admin/service/iface/JdGoodsService.java
#	com-platform-admin-service/src/main/java/com/platform/admin/service/AdminServiceApplication.java
#	com-platform-admin-service/src/main/java/com/platform/admin/service/api/JdAPI.java
#	com-platform-admin-service/src/main/java/com/platform/admin/service/biz/JdGoodsSyncBiz.java
#	com-platform-admin-service/src/main/java/com/platform/admin/service/mapper/JdGoodsMapper.java
#	com-platform-admin-service/src/main/java/com/platform/admin/service/service/JdGoodsServiceController.java
#	com-platform-admin-service/src/main/resources/application.properties
#	com-platform-common/src/main/java/com/platform/common/config/EsConfig.java
#	com-platform-common/src/main/java/com/platform/common/contanst/EsConstanst.java
#	com-platform-common/src/main/java/com/platform/common/modal/GoodsInfo.java
#	com-platform-common/src/main/java/com/platform/common/modal/PageInfo.java
#	com-platform-common/src/main/java/com/platform/common/util/es/EsWriteUtil.java
#	com-platform-common/src/main/java/com/platform/common/util/es/RestClientUtil.java
#	com-platform-front-service/src/main/java/com/platform/front/service/api/TaobaoGoodsAPI.java
#	com-platform-front-service/src/main/java/com/platform/front/service/biz/JdGoodsQueryBiz.java
#	com-platform-front-service/src/main/resources/application.properties
#	com-platform-front-web/src/main/resources/static/js/list.js
#	com-platform-task/pom.xml
#	com-platform-task/src/main/java/com/platform/task/biz/JdGoodsBiz.java
#	com-platform-task/src/main/java/com/platform/task/job/JdGoodsSyncJob.java
#	com-platform-task/src/main/resources/application.properties
#	doc/goodsinfo.txt

+++ com-platform-common/src/main/java/com/platform/common/modal/PageInfo.java
@@ -11,10 +11,10 @@ import java.io.Serializable;
 @Getter
 @Setter
 public class PageInfo  implements Serializable {
-    private long pageSize = 25;
-    private long pageIndex = 1;
+    private int pageSize = 25;
+    private int pageIndex = 1;
 
-    public long getPageCount(long totalCount) {
+    public int getPageCount(int totalCount) {
         return (totalCount  +  pageSize  - 1) / pageSize;
     }
 }
